Move JVM session creation into SDK via randomUUID call.
Notes:

Follows patterns from Node/Python to create session using UUID call
Intention is to keep customer interface the same. For single completions this is trivial, for sessions this means storing the metadata on the session class and passing it to record from there.
A bunch of tests broke because we assert on the # of calls to the mock. A lot of the changeset is fixing these.

Co-authored-by: Nico Tonozzi <nico@freeplay.ai>